Extract shared dashboard mission model
